Warriors can now be deselected by pressing the escape key.
Fixed the height of the layouts. Improved some listeners. New known
issue: even though the cells are freed from their objects when the
enemies advance, the listeners still think all the freed cells are
taken. The tester says that the cells are correctly freed. However, even
using the same functions, this is not the case when the PlacementLayout
checks the cell through the Game and the Map in the logic.